requirements.txt
daphne
channels
channels_redis
redis
jango